% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waxer.R
\docType{package}
\name{waxer-package}
\alias{waxer}
\alias{waxer-package}
\title{waxer: Wikimedia Analytics Query Service (AQS) API}
\description{
This is an R wrapper for the
\href{https://www.mediawiki.org/wiki/Wikidata_query_service}{Wikimedia Analytics Query Service (AQS)}.
The \href{https://wikimedia.org/api/rest_v1/}{REST API} provides cacheable and
straightforward access to Wikimedia content and data, in machine-readable
formats. This particular wrapper is for the \verb{/metrics} endpoint.
}
\section{Definitions}{

\subsection{Editor type}{

The \code{editor_type} parameter in functions allows you to filter the type of
editors having performed the related action. Can be \code{anonymous} for users
not logged in,\verb{ bot group} for logged in users that are part of the bot
group, \code{bot} for logged in users whose name contains \code{bot} (high probability
of being a bot, even if counter examples exists), and\code{ user} for logged-in
users not in \verb{bot group} nor \code{bot} sets. Finally, you can use \code{all} to not
filter by editor type.
}

\subsection{Page type}{

The \code{page_type} parameter in functions allows you to filter for the type of
page over which the action is performed. Can be \code{content} for pages
belonging in content namespaces, also known as "main" namespaces. This page
type is also referred to as articles, and for most wikis includes pages in
namespace 0 only. Can also be \code{non-content} for pages in namespaces not
considered content (talk pages, user pages etc). Finally, you can use \code{all}
to not filter by page type.
}
}

\section{Wikimedia projects}{

The package comes bundled with a list of \emph{active} Wikibooks, Wikinews,
Wikipedia, Wikiquote, Wikisource, Wikiversity, Wikivoyage, and Wiktionary
wikis.\preformatted{wikis <- read.csv(system.file("extdata/wikis.csv", package = "waxer"))
}

The columns are:
\describe{
\item{\code{language}}{e.g. "English", "Arabic"}
\item{\code{language_code}}{e.g. "en", "ar"}
\item{\code{project_code}}{e.g. "wikipedia", "wikibooks"}
\item{\code{url}}{e.g. "en.wikipedia.org", "ar.wikibooks.org"}
\item{\code{direction}}{e.g. "left-to-right", "right-to-left"}
}

This dataset was generated from the
\href{https://www.mediawiki.org/wiki/Special:SiteMatrix}{Wikimedia site matrix}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bearloga/waxer}
  \item Report bugs at \url{https://github.com/bearloga/waxer/issues}
}

}
\author{
\strong{Maintainer}: Mikhail Popov \email{mikhail@wikimedia.org}

Other contributors:
\itemize{
  \item Wikimedia Foundation [copyright holder]
}

}
\keyword{internal}
