% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{wx_edited_pages}
\alias{wx_edited_pages}
\title{Edited pages counts}
\usage{
wx_edited_pages(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  activity_level = c("all", "1-4", "5-24", "25-99", "100+"),
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{editor_type}{If you want to filter by editor-type, use one of
"anonymous", "group-bot" (registered accounts belonging to the bot group),
"name-bot" (registered accounts not belonging to the bot group but having
bot-like names) or "user" (registered account not in bot group nor having
bot-like name). If you are interested in edits regardless of their editor
type, use "all-editor-types" (default).}

\item{page_type}{If you want to filter by page-type, use one of: "content"
(edits made in content namespaces) or "non-content" (edits made in
non-content namespaces). If you are interested in editors regardless of
their page type, use "all" (default).}

\item{activity_level}{If you want to filter by activity-level (number of
edits), use one of: "1-4", "5-24", "25-99", or "100+" edits. If you are
interested in pages regardless of their activity-level, use
"all" (default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\description{
Number of edited pages for a given project and timespan.
}
\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
# Monthly edited pages on Hindi Wikipedia:
wx_edited_pages("hi.wikipedia", page_type = "content", granularity = "monthly")
}
