% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediareqs.R
\name{wx_mediareqs_file}
\alias{wx_mediareqs_file}
\title{Mediarequest counts for a media file}
\usage{
wx_mediareqs_file(
  file_path,
  referer,
  media_type = c("all", "image", "video", "audio", "document", "other"),
  agent_type = c("all", "user", "bot"),
  granularity = c("daily", "monthly"),
  start_date = "20191201",
  end_date = "20200101"
)
}
\arguments{
\item{file_path}{One or more paths to the file on \code{upload.wikimedia.org},
which is the file storage for all media in any wiki.}

\item{referer}{The place that the request was made from. It can be any
Wikimedia project (e.g. de.wikipedia), "all", "internal", "external",
"search-engine", "unknown" or "none".}

\item{media_type}{The media type that each file belongs to. It can be
image, audio, video, document, or other. Default is "all" media types.
See Media section below.}

\item{agent_type}{If you want to filter by agent type, use "user", "bot"/"spider",
or "automated" (refer to \href{https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Traffic/BotDetection}{wikitech:Analytics/Data Lake/Traffic/BotDetection}).
If you are interested in pageviews regardless of agent type, use "all"
(default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{referer}}{referer}
\item{\code{file_path}}{media file path provided by the user}
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{requests}}{number of requests}
}
}
\description{
Daily and monthly counts of mediarequests for each media file
stored in the wiki servers (as long as the count is higher than one).
}
\section{Example usage}{

The example below retrieves the request counts for
\href{https://commons.wikimedia.org/wiki/File:Hadley-wickham2016-02-04.jpg}{File:Hadley-wickham2016-02-04.jpg}
which is used on the English Wikipedia article
\href{https://en.wikipedia.org/wiki/Hadley_Wickham}{\emph{Hadley Wickham}}
as the main image in the infobox.
}

\section{Media}{

File types are obtained by parsing the file extension, and then classified
according to the following table:\tabular{ll}{
   Extensions \tab Media type \cr
   svg, png, tiff, tiff, jpeg, jpeg, gif, xcf, webp, bmp \tab image \cr
   mp3, ogg, oga, flac, wav, midi, midi \tab audio \cr
   webm, ogv \tab video \cr
   pdf, djvu, srt, txt \tab document \cr
   (all other extensions) \tab other \cr
}
}

\section{Limitations}{

\itemize{
\item The ability of splitting and filtering by referrer is limited to data
from May 2019 onward. Before that, referrer is only split in internal,
external, and unknown.
\item The beginning of mediarequest data is the 1st of January 2015.
\item The ability of splitting and filtering by agent type (user, spider) is
limited to data from May 2019 onward.
\item About 0.7\% of mediarequests are prefetches coming from Media Viewer
(more details in \href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Media_metrics}{wikitech:Analytics/AQS/Media_metrics})
}
}

\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
wx_mediareqs_file(
  "/wikipedia/commons/f/fa/Hadley-wickham2016-02-04.jpg", "en.wikipedia",
  agent_type = "user", granularity = "monthly",
  start_date = "20200101", end_date = "20200301"
)
}
\seealso{
Other {content data & metrics}: 
\code{\link{wx_edited_pages}()},
\code{\link{wx_mediareqs_referer}()},
\code{\link{wx_most_requested_files}()},
\code{\link{wx_new_pages}()},
\code{\link{wx_page_edits}()},
\code{\link{wx_project_edits}()},
\code{\link{wx_top_edited_pages}()},
\code{\link{wx_total_pages}()}

Other {mediarequests data & metrics}: 
\code{\link{wx_mediareqs_referer}()},
\code{\link{wx_most_requested_files}()}
}
\concept{{content data & metrics}}
\concept{{mediarequests data & metrics}}
