% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediareqs.R
\name{wx_mediareqs_referer}
\alias{wx_mediareqs_referer}
\title{Mediarequest counts per referer}
\usage{
wx_mediareqs_referer(
  referer,
  media_type = c("all", "image", "video", "audio", "document", "other"),
  agent_type = c("all", "user", "bot"),
  granularity = c("daily", "monthly"),
  start_date = "20191201",
  end_date = "20200101"
)
}
\arguments{
\item{referer}{The place that the request was made from. It can be any
Wikimedia project (e.g. de.wikipedia), "all", "internal", "external",
"search-engine", "unknown" or "none".}

\item{media_type}{The media type that each file belongs to. It can be
image, audio, video, document, or other. Default is "all" media types.
See Media section below.}

\item{agent_type}{If you want to filter by agent type, use "user", "bot"/"spider",
or "automated" (refer to \href{https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Traffic/BotDetection}{wikitech:Analytics/Data Lake/Traffic/BotDetection}).
If you are interested in pageviews regardless of agent type, use "all"
(default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{project}}{project}
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{requests}}{number of media file requests}
}
}
\description{
Daily and monthly aggregation of hits to media files, split by
the referrer Referrer means either external, internal, unknown, or any
Wikimedia wiki that the resource was requested from.
}
\section{Media}{

File types are obtained by parsing the file extension, and then classified
according to the following table:\tabular{ll}{
   Extensions \tab Media type \cr
   svg, png, tiff, tiff, jpeg, jpeg, gif, xcf, webp, bmp \tab image \cr
   mp3, ogg, oga, flac, wav, midi, midi \tab audio \cr
   webm, ogv \tab video \cr
   pdf, djvu, srt, txt \tab document \cr
   (all other extensions) \tab other \cr
}
}

\section{Limitations}{

\itemize{
\item The ability of splitting and filtering by referrer is limited to data
from May 2019 onward. Before that, referrer is only split in internal,
external, and unknown.
\item The beginning of mediarequest data is the 1st of January 2015.
\item The ability of splitting and filtering by agent type (user, spider) is
limited to data from May 2019 onward.
\item About 0.7\% of mediarequests are prefetches coming from Media Viewer
(more details in \href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Media_metrics}{wikitech:Analytics/AQS/Media_metrics})
}
}

\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
wx_mediareqs_referer("de.wikipedia", granularity = "monthly")
}
\seealso{
\itemize{
\item \href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Mediarequests}{wikitech:Analytics/AQS/Mediarequests}
}

Other {content data & metrics}: 
\code{\link{wx_edited_pages}()},
\code{\link{wx_mediareqs_file}()},
\code{\link{wx_most_requested_files}()},
\code{\link{wx_new_pages}()},
\code{\link{wx_page_edits}()},
\code{\link{wx_project_edits}()},
\code{\link{wx_top_edited_pages}()},
\code{\link{wx_total_pages}()}

Other {mediarequests data & metrics}: 
\code{\link{wx_mediareqs_file}()},
\code{\link{wx_most_requested_files}()}
}
\concept{{content data & metrics}}
\concept{{mediarequests data & metrics}}
