% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{wx_mediawiki_api}
\alias{wx_mediawiki_api}
\title{Generate a rate-limited MediaWiki API caller}
\usage{
wx_mediawiki_api(project)
}
\arguments{
\item{project}{The project for which to generate queryable MW API endpoint}
}
\value{
A rate-limited function with parameters:
\describe{
\item{\code{query}}{query after endpoint; e.g. \verb{action=query&format=json&titles=Main\%20page}}
}
}
\description{
Generates a function which makes a request to MediaWiki APIs.
This generator is primarily intended for internal use, but can
theoretically be used by others to write custom MediaWiki API queries. The
generated function is rate-limited at 100 queries per second.
}
\seealso{
\itemize{
\item \href{https://www.mediawiki.org/wiki/API:Main_page}{MediaWiki API}
}
}
