% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediareqs.R
\name{wx_most_requested_files}
\alias{wx_most_requested_files}
\title{Most requested files for a referer}
\usage{
wx_most_requested_files(
  referer,
  media_type = c("all", "image", "video", "audio", "document", "other"),
  granularity = c("daily", "monthly"),
  start_date = "20191231",
  end_date = "20200101"
)
}
\arguments{
\item{referer}{The place that the request was made from. It can be any
Wikimedia project (e.g. de.wikipedia), "all", "internal", "external",
"search-engine", "unknown" or "none".}

\item{media_type}{The media type that each file belongs to. It can be
image, audio, video, document, or other. Default is "all" media types.
See Media section below.}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{file_path}}{the media file's path on upload.wikimedia.org}
\item{\code{requests}}{number of requests}
\item{\code{rank}}{\code{requests}-based ranking}
}
}
\description{
The most requested media files per referer and per media type.
}
\section{Media}{

File types are obtained by parsing the file extension, and then classified
according to the following table:\tabular{ll}{
   Extensions \tab Media type \cr
   svg, png, tiff, tiff, jpeg, jpeg, gif, xcf, webp, bmp \tab image \cr
   mp3, ogg, oga, flac, wav, midi, midi \tab audio \cr
   webm, ogv \tab video \cr
   pdf, djvu, srt, txt \tab document \cr
   (all other extensions) \tab other \cr
}
}

\section{Limitations}{

\itemize{
\item The ability of splitting and filtering by referrer is limited to data
from May 2019 onward. Before that, referrer is only split in internal,
external, and unknown.
\item The beginning of mediarequest data is the 1st of January 2015.
\item The ability of splitting and filtering by agent type (user, spider) is
limited to data from May 2019 onward.
\item About 0.7\% of mediarequests are prefetches coming from Media Viewer
(more details in \href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Media_metrics}{wikitech:Analytics/AQS/Media_metrics})
}
}

\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
wx_most_requested_files("en.wikipedia", media_type = "video", granularity = "monthly")
}
\seealso{
Other {content data & metrics}: 
\code{\link{wx_edited_pages}()},
\code{\link{wx_mediareqs_file}()},
\code{\link{wx_mediareqs_referer}()},
\code{\link{wx_new_pages}()},
\code{\link{wx_page_edits}()},
\code{\link{wx_project_edits}()},
\code{\link{wx_top_edited_pages}()},
\code{\link{wx_total_pages}()}

Other {mediarequests data & metrics}: 
\code{\link{wx_mediareqs_file}()},
\code{\link{wx_mediareqs_referer}()}
}
\concept{{content data & metrics}}
\concept{{mediarequests data & metrics}}
