% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{wx_query_api}
\alias{wx_query_api}
\title{Make HTTP request and process response}
\usage{
wx_query_api(reqs_per_second = 200)
}
\arguments{
\item{reqs_per_second}{Maximum requests allowed per second (default 200)}
}
\description{
Makes a request to
\href{https://wikimedia.org/api/rest_v1/}{wikimedia.org/api/rest_v1/}.
}
