% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{wx_top_edited_pages}
\alias{wx_top_edited_pages}
\title{Top edited pages (by edits)}
\usage{
wx_top_edited_pages(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  granularity = c("daily", "monthly"),
  start_date = "20191201",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{editor_type}{If you want to filter by editor-type, use one of
"anonymous", "group-bot" (registered accounts belonging to the bot group),
"name-bot" (registered accounts not belonging to the bot group but having
bot-like names) or "user" (registered account not in bot group nor having
bot-like name). If you are interested in edits regardless of their editor
type, use "all-editor-types" (default).}

\item{page_type}{If you want to filter by page-type, use one of: "content"
(edits made in content namespaces) or "non-content" (edits made in
non-content namespaces). If you are interested in editors regardless of
their page type, use "all" (default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{page_name}}{page name; \strong{NOTE}: in wikitext, the first letter
of the target page is automatically capitalized}
\item{\code{edits}}{number of edits}
\item{\code{rank}}{\code{edits}-based ranking}
}
}
\description{
Lists the 100 most edited pages for a given project.
}
\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
# Top edited articles on German Wikivoyage:
wx_top_edited_pages("de.wikivoyage", page_type = "content")
}
\seealso{
\link{wx_top_viewed_pages}

Other {content data & metrics}: 
\code{\link{wx_edited_pages}()},
\code{\link{wx_mediareqs_file}()},
\code{\link{wx_mediareqs_referer}()},
\code{\link{wx_most_requested_files}()},
\code{\link{wx_new_pages}()},
\code{\link{wx_page_edits}()},
\code{\link{wx_project_edits}()},
\code{\link{wx_total_pages}()}
}
\concept{{content data & metrics}}
