% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{wx_top_editors}
\alias{wx_top_editors}
\title{Top editors by edit count}
\usage{
wx_top_editors(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  start_date = "20191231",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{editor_type}{If you want to filter by editor-type, use one of
"anonymous", "group-bot" (registered accounts belonging to the bot group),
"name-bot" (registered accounts not belonging to the bot group but having
bot-like names) or "user" (registered account not in bot group nor having
bot-like name). If you are interested in edits regardless of their editor
type, use "all-editor-types" (default).}

\item{page_type}{If you want to filter by page-type, use one of: "content"
(edits made in content namespaces) or "non-content" (edits made in
non-content namespaces). If you are interested in editors regardless of
their page type, use "all" (default).}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{project}}{project}
\item{\code{date}}{\code{Date}}
\item{\code{user_text}}{the editor's username; \code{NA} if anonymous (instead of IP address)}
\item{\code{edits}}{number of edits the user made on \code{date}}
\item{\code{rank}}{the rank based on \code{edits}}
}
}
\description{
Top 100 editors per day by number of edits. \strong{Caution}:
anonymous editors IPs are replaced by \code{NA}.
}
\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
\dontrun{
wx_top_editors(
  "commons.wikimedia",
  editor_type = "bot",
  page_type = "content",
  start_date = "20191001",
  end_date = "20191031"
)
}
}
\seealso{
Other {data & metrics on users}: 
\code{\link{wx_active_editors}()},
\code{\link{wx_newly_registered_users}()}
}
\concept{{data & metrics on users}}
