% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{wx_top_viewed_articles}
\alias{wx_top_viewed_articles}
\title{Top viewed articles}
\usage{
wx_top_viewed_articles(
  project = "mediawiki",
  access_method = c("all", "desktop", "mobile web", "mobile app"),
  granularity = c("daily", "monthly"),
  start_date = "20191231",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{access_method}{If you want to filter by access method, use one of:
"desktop", "mobile app", or "mobile web". If you are interested in
pageviews regardless of access method, use "all" (default).}

\item{granularity}{The time unit for the response data; "daily" by default
but if you want to get the top articles of a whole month use "monthly".}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{article}}{article name; \strong{NOTE}: in wikitext, the first letter
of the target page is automatically capitalized and spaces are
represented as underscores}
\item{\code{views}}{number of page-views (see
\href{https://meta.wikimedia.org/wiki/Research:Page_view}{meta:Research:Page view}
for more information on what is considered a page-view)}
\item{\code{rank}}{\code{views}-based ranking}
}
}
\description{
Lists the 1000 most viewed articles for a given project and
timespan (month or day).
}
