% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{wx_top_viewed_pages}
\alias{wx_top_viewed_pages}
\title{Top viewed pages}
\usage{
wx_top_viewed_pages(
  project,
  access_method = c("all", "desktop", "mobile web", "mobile app"),
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20191231"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{access_method}{If you want to filter by access method, use one of:
"desktop", "mobile app", or "mobile web". If you are interested in
pageviews regardless of access method, use "all" (default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{page_name}}{page name; \strong{NOTE}: in wikitext, the first letter
of the target page is automatically capitalized}
\item{\code{views}}{number of page-views (see
\href{https://meta.wikimedia.org/wiki/Research:Page_view}{meta:Research:Page view}
for more information on what is considered a page-view)}
\item{\code{rank}}{\code{views}-based ranking}
}
}
\description{
Lists the 1000 most viewed pages for a given project and
timespan (month or day).
}
\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
wx_top_viewed_pages("en.wikipedia")
}
\seealso{
Other {traffic data & metrics}: 
\code{\link{wx_page_views}()},
\code{\link{wx_project_views}()},
\code{\link{wx_unique_devices}()}
}
\concept{{traffic data & metrics}}
