% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{wx_total_pages}
\alias{wx_total_pages}
\title{Total pages counts}
\usage{
wx_total_pages(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{editor_type}{If you want to filter by editor-type, use one of
"anonymous", "group-bot" (registered accounts belonging to the bot group),
"name-bot" (registered accounts not belonging to the bot group but having
bot-like names) or "user" (registered account not in bot group nor having
bot-like name). If you are interested in edits regardless of their editor
type, use "all-editor-types" (default).}

\item{page_type}{If you want to filter by page-type, use one of: "content"
(edits made in content namespaces) or "non-content" (edits made in
non-content namespaces). If you are interested in editors regardless of
their page type, use "all" (default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{project}}{project}
\item{\code{date}}{\code{Date}}
\item{\code{total_pages}}{Number of total (new + existing) pages},
\item{\code{existing_pages}}{Number of pages already in project},
\item{\code{new_pages}}{Number of new pages created}
}
}
\description{
Number of total pages for a given project and timespan. This
is a convenience function that combines \link{wx_new_pages} and \link[base:cumsum]{base::cumsum}.
}
\details{
Since there is no endpoint for this metric and it must be derived from the
new pages metric by taking its cumulative value, this function calls
\link{wx_new_pages} with \code{start_date = 2000-01-01} so that it can correctly
calculate cumulative number of pages from the very beginning of \code{project}.
Then it crops the results to start with the actual provided \code{start_date}.
}
\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
\dontrun{
wx_total_pages("en.wikipedia")
}
}
\seealso{
Other {content data & metrics}: 
\code{\link{wx_edited_pages}()},
\code{\link{wx_mediareqs_file}()},
\code{\link{wx_mediareqs_referer}()},
\code{\link{wx_most_requested_files}()},
\code{\link{wx_new_pages}()},
\code{\link{wx_page_edits}()},
\code{\link{wx_project_edits}()},
\code{\link{wx_top_edited_pages}()}
}
\concept{{content data & metrics}}
