% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{wx_unique_devices}
\alias{wx_unique_devices}
\title{Unique device counts}
\usage{
wx_unique_devices(
  project,
  access_site = c("all", "desktop", "mobile"),
  granularity = c("daily", "monthly"),
  start_date = "20191231",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{access_site}{If you want to filter by accessed site, use "desktop"
or "mobile". If you are interested in unique devices regardless of
accessed site, use "all" (default). \strong{\emph{NOT}} to be confused with
\code{access_method} parameter in other functions.}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{date}}{\code{Date}; beginning of each month if \code{granularity = "monthly"}}
\item{\code{devices}}{Estimated number of unique devices.}
\item{\code{offset}}{The number added to \code{underestimate} to produce \code{devices} estimate.}
\item{\code{underestimate}}{The count produced by WMF-Last-Access cookie counting, which, by definition, underreports users as we will not be counting users with a fresh session or users browsing without cookies.}
}
}
\description{
Unique Devices dataset (v1) as developed and maintained by the
Wikimedia Foundation since January 1st, 2016. This data set contains the
number of unique devices having visited Wikipedia or its sisters project
over some period of time. The Analytics team counts unique devices per
project per day and month in a way that does not uniquely identify,
fingerprint or otherwise track users.
}
\section{Site}{

Desktop devices can (and \emph{do}) access the mobile website, usually because
the user clicked on a link shared by a user from a mobile device. So do not
use this metric as a proxy for mobile vs desktop devices.

To learn more about how mobile traffic is handled by Wikimedia, refer to
\href{https://www.mediawiki.org/wiki/Reading/Web/Mobile}{mw:Reading/Web/Mobile}.
}

\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
\dontrun{
wx_unique_devices("en.wikipedia")
}
}
\seealso{
\itemize{
\item \href{https://meta.wikimedia.org/wiki/Research:Unique_devices}{wikitech:Research:Unique devices}
\item \href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Unique_Devices}{wikitech:Analytics/AQS/Unique Devices}
\item \href{https://blog.wikimedia.org/2016/03/30/unique-devices-dataset/}{Introducing the unique devices dataset: a new way to estimate reach on Wikimedia projects}
\item \href{https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Traffic/Unique_Devices/Last_access_solution}{wikitech:Analytics/Data Lake/Traffic/Unique Devices/Last access solution}
}
}
