% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waxer.R
\docType{package}
\name{waxer-package}
\alias{waxer}
\alias{waxer-package}
\title{waxer: Wikimedia Analytics Query Service (AQS) API}
\description{
This is an R wrapper for the
\href{https://www.mediawiki.org/wiki/Wikidata_query_service}{Wikimedia Analytics Query Service (AQS)}.
The \href{https://wikimedia.org/api/rest_v1/}{REST API} provides cacheable and
straightforward access to Wikimedia content and data, in machine-readable
formats. This particular wrapper is for the \verb{/metrics} endpoint.
}
\section{Wikimedia projects}{

The package comes bundled with a list of \emph{active} Wikibooks, Wikinews,
Wikipedia, Wikiquote, Wikisource, Wikiversity, Wikivoyage, and Wiktionary
wikis.\preformatted{wikis <- read.csv(system.file("extdata/wikis.csv", package = "waxer"))
}

The columns are:
\describe{
\item{\code{language}}{e.g. "English", "Arabic"}
\item{\code{language_code}}{e.g. "en", "ar"}
\item{\code{project_code}}{e.g. "wikipedia", "wikibooks"}
\item{\code{url}}{e.g. "en.wikipedia.org", "ar.wikibooks.org"}
\item{\code{direction}}{e.g. "left-to-right", "right-to-left"}
}

This dataset was generated from the
\href{https://www.mediawiki.org/wiki/Special:SiteMatrix}{Wikimedia site matrix}.
}

\section{Metrics}{

\subsection{Traffic}{

\describe{
\item{\link{wx_top_viewed_articles}}{Top 1000 viewed articles}
\item{\link{wx_unique_device}}{Unique devices}
}
}

\subsection{Editors}{

\describe{
\item{\link{wx_active_editors}}{Number of active editors}
\item{\link{wx_top_editors}}{Top 100 editors per day by number of edits}
}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bearloga/waxer}
  \item Report bugs at \url{https://github.com/bearloga/waxer/issues}
}

}
\author{
\strong{Maintainer}: Mikhail Popov \email{mikhail@wikimedia.org}

Other contributors:
\itemize{
  \item Wikimedia Foundation [copyright holder]
}

}
\keyword{internal}
