% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{wx_active_editors}
\alias{wx_active_editors}
\title{Active editors counts}
\usage{
wx_active_editors(
  project,
  editor_type = c("all", "anonymous", "user", "bot", "bot group"),
  page_type = c("all", "content", "non-content"),
  activity_level = c("all", "1-4", "5-24", "25-99", "100+"),
  granularity = c("daily", "monthly"),
  start_date = "20191101",
  end_date = "20191231"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{editor_type}{If you want to filter by editor-type, use one of
"anonymous", "group-bot" (registered accounts belonging to the bot group),
"name-bot" (registered accounts not belonging to the bot group but having
bot-like names) or "user" (registered account not in bot group nor having
bot-like name). If you are interested in edits regardless of their editor
type, use "all-editor-types" (default).}

\item{page_type}{If you want to filter by page-type, use one of: "content"
(edits made in content namespaces) or "non-content" (edits made in
non-content namespaces). If you are interested in editors regardless of
their page type, use "all" (default).}

\item{activity_level}{If you want to filter by activity-level (number of
edits), use one of: "1-4", "5-24", "25-99", or "100+" edits. If you are
interested in editors regardless of their activity-level, use
"all" (default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default) and monthly.}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with columns
\describe{
\item{\code{project}}{project}
\item{\code{date}}{\code{Date}; beginning of the month if \code{granularity = "monthly"}}
\item{\code{editors}}{Number of active editors}
}
}
\description{
Number of active editors. \strong{Caution}: This metric measures
editors' activity, therefore it includes edits on redirects.
}
\section{Granularity and dates}{

For "monthly" \code{granularity}, the \code{start_date} and \code{end_date} need to contain
the full month.

For example, \code{start_date = "20191101"} and
\code{end_date = "20191231"} would only give the monthly total for 2019-11. To
include 2019-12, use \code{end_date = "20200101"}.

Frustratingly, \code{start_date = "20191201"} and \code{end_date = "20191231"} does
\strong{\emph{not}} yield 2019-12 monthly total. Use \code{end_date = "20200101"} for that.
}

\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
wx_active_editors(
  "commons.wikimedia",
  editor_type = "bot",
  page_type = "content",
  activity_level = "100+",
  granularity = "monthly",
  start_date = "20200101",
  end_date = "20200301"
)
}
\seealso{
\href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Wikistats_2}{wikitech:Analytics/AQS/Wikistats 2}

Other {data & metrics on users}: 
\code{\link{wx_newly_registered_users}()},
\code{\link{wx_top_editors}()}
}
\concept{{data & metrics on users}}
