% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{wx_project_views}
\alias{wx_project_views}
\title{Project page-view counts}
\usage{
wx_project_views(
  project,
  access_method = c("all", "desktop", "mobile web", "mobile app"),
  agent_type = c("all", "user", "bot"),
  granularity = c("daily", "monthly", "hourly"),
  start_date = "20191231",
  end_date = "20200101"
)
}
\arguments{
\item{project}{The name of any Wikimedia project formatted like
\verb{\{language code\}.\{project name\}}, for example en.wikipedia. You may pass
en.wikipedia.org and the .org will be stripped off. For projects like
Wikimedia Commons (without language codes), use commons.wikimedia. For
projects like www.mediawiki.org, you can use that full string, or just use
mediawiki or mediawiki.org.}

\item{access_method}{If you want to filter by access method, use one of:
"desktop", "mobile app", or "mobile web". If you are interested in
pageviews regardless of access method, use "all" (default).}

\item{agent_type}{If you want to filter by agent type, use "user" or "bot".
If you are interested in pageviews regardless of agent type, use "all"
(default).}

\item{granularity}{The time unit for the response data. As of today,
supported values are daily (default), monthly, and hourly. In case of
hourly granularity, for simplicity, the function retrieves data from the
very start of \code{start_date} (hour 00) to the very end of \code{end_date} (hour
23).}

\item{start_date}{The date of the first day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}

\item{end_date}{The date of the last day to include, in YYYYMMDD format.
Can also be a \code{Date} or a \code{POSIXt} object, which will be auto-formatted.}
}
\value{
A tibble data frame with the following columns:
\describe{
\item{\code{project}}{project}
\item{\code{date} if daily or monthly granularity, \code{time} if hourly granularity}{\code{Date}/\code{POSIXct}; all times are in UTC\verb{\} \\item\{}views`}{total number of page-views for the project}
}
}
\description{
Project page-view counts
}
\section{License}{

Data retrieved from the API endpoint is available under the
\href{https://creativecommons.org/publicdomain/zero/1.0/}{CC0 1.0 license}.
}

\examples{
# block from 2017-04-29 to 2020-01-15
wx_project_views(
  "tr.wikipedia", # Turkish
  granularity = "monthly",
  start_date = "20191001",
  end_date = "20200229"
)
}
\seealso{
\itemize{
\item \href{https://meta.wikimedia.org/wiki/Research:Page_view}{meta:Reseach:Page view}
\item \href{https://wikitech.wikimedia.org/wiki/Analytics/AQS/Pageviews}{wikitech:Analytics/AQS/Pageviews}
}

Other {traffic data & metrics}: 
\code{\link{wx_page_views}()},
\code{\link{wx_top_viewed_pages}()},
\code{\link{wx_unique_devices}()}
}
\concept{{traffic data & metrics}}
